backend/requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16

boto3>=1.18.0,<1.19.0  # For AWS S3 interactions
django-storages>=1.11.1,<1.12.0  # For handling storage backends like S3